pi pico w/requirements.txt.txt
micropython-tm1637
micropython-urequests
micropython-network
micropython-max6675
